[breaking] initial changes for Spectrum refactor
